Refactor discord events path
